clean up directory structure by renaming and moving files
