+++ frontend/src/pages/employer/EmployerPosts.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import SearchBar from "../components/SearchBar";
+import SearchBar from "../../components/SearchBar";
 
 function EmployerPosts() {
     const jobs = [
